refactor: 통합 테스트도 별도 패키지로 분리
