+++ src/test/java/com/techfork/domain/activity/controller/ActivityControllerIntegrationTest.java
@@ -1,7 +1,6 @@
 package com.techfork.domain.activity.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.techfork.domain.activity.bookmark.dto.BookmarkRequest;
 import com.techfork.domain.activity.dto.ReadPostRequest;
 import com.techfork.domain.activity.dto.SearchHistoryRequest;
 import com.techfork.domain.activity.entity.ReadPost;
@@ -310,180 +309,6 @@ class ActivityControllerIntegrationTest extends IntegrationTestBase {
         assertThat(searchHistories).hasSize(3);
     }
 
-    // ===== 북마크 추가 테스트 =====
-
-    @Test
-    @DisplayName("북마크 추가 성공")
-    void addBookmark_Success() throws Exception {
-        // Given
-        BookmarkRequest request = new BookmarkRequest(testPost1.getId());
-
-        // When & Then
-        mockMvc.perform(post("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request)))
-                .andDo(print())
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.isSuccess").value(true));
-
-        // DB 검증
-        List<Bookmark> bookmarks = bookmarkRepository.findAll();
-        assertThat(bookmarks).hasSize(1);
-        assertThat(bookmarks.get(0).getUser().getId()).isEqualTo(testUser.getId());
-        assertThat(bookmarks.get(0).getPost().getId()).isEqualTo(testPost1.getId());
-    }
-
-    @Test
-    @DisplayName("북마크 추가 실패 - 이미 북마크한 게시글")
-    void addBookmark_Fail_AlreadyExists() throws Exception {
-        // Given - 이미 북마크 생성
-        Bookmark existingBookmark = Bookmark.create(testUser, testPost1, LocalDateTime.now());
-        bookmarkRepository.save(existingBookmark);
-
-        BookmarkRequest request = new BookmarkRequest(testPost1.getId());
-
-        // When & Then
-        mockMvc.perform(post("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request)))
-                .andDo(print())
-                .andExpect(status().isConflict())
-                .andExpect(jsonPath("$.isSuccess").value(false))
-                .andExpect(jsonPath("$.code").value("BOOKMARK409_1"));
-    }
-
-    @Test
-    @DisplayName("북마크 추가 실패 - 존재하지 않는 게시글")
-    void addBookmark_Fail_PostNotFound() throws Exception {
-        // Given
-        BookmarkRequest request = new BookmarkRequest(99999L);
-
-        // When & Then
-        mockMvc.perform(post("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request)))
-                .andDo(print())
-                .andExpect(status().isNotFound())
-                .andExpect(jsonPath("$.isSuccess").value(false));
-    }
-
-    // ===== 북마크 삭제 테스트 =====
-
-    @Test
-    @DisplayName("북마크 삭제 성공")
-    void deleteBookmark_Success() throws Exception {
-        // Given - 북마크 생성
-        Bookmark bookmark = Bookmark.create(testUser, testPost1, LocalDateTime.now());
-        bookmarkRepository.save(bookmark);
-
-        BookmarkRequest request = new BookmarkRequest(testPost1.getId());
-
-        // When & Then
-        mockMvc.perform(delete("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request)))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.isSuccess").value(true));
-
-        // DB 검증 - 삭제되었는지 확인
-        List<Bookmark> bookmarks = bookmarkRepository.findAll();
-        assertThat(bookmarks).isEmpty();
-    }
-
-    @Test
-    @DisplayName("북마크 삭제 실패 - 북마크가 존재하지 않음")
-    void deleteBookmark_Fail_NotFound() throws Exception {
-        // Given - 북마크 없음
-        BookmarkRequest request = new BookmarkRequest(testPost1.getId());
-
-        // When & Then
-        mockMvc.perform(delete("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request)))
-                .andDo(print())
-                .andExpect(status().isNotFound())
-                .andExpect(jsonPath("$.isSuccess").value(false))
-                .andExpect(jsonPath("$.code").value("BOOKMARK404_1"));
-    }
-
-    // ===== 북마크 목록 조회 테스트 =====
-
-    @Test
-    @DisplayName("북마크 목록 조회 성공 - 빈 목록")
-    void getBookmarks_Success_Empty() throws Exception {
-        // When & Then
-        mockMvc.perform(get("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .param("size", "20"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.isSuccess").value(true))
-                .andExpect(jsonPath("$.data.bookmarks").isArray())
-                .andExpect(jsonPath("$.data.bookmarks").isEmpty())
-                .andExpect(jsonPath("$.data.hasNext").value(false));
-    }
-
-    @Test
-    @DisplayName("북마크 목록 조회 성공 - 여러 개")
-    void getBookmarks_Success_Multiple() throws Exception {
-        // Given - 여러 개의 북마크 생성
-        Bookmark bookmark1 = Bookmark.create(testUser, testPost1, LocalDateTime.now().minusHours(1));
-        Bookmark bookmark2 = Bookmark.create(testUser, testPost2, LocalDateTime.now());
-        bookmarkRepository.save(bookmark1);
-        bookmarkRepository.save(bookmark2);
-
-        // When & Then
-        mockMvc.perform(get("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .param("size", "20"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.isSuccess").value(true))
-                .andExpect(jsonPath("$.data.bookmarks").isArray())
-                .andExpect(jsonPath("$.data.bookmarks.length()").value(2))
-                .andExpect(jsonPath("$.data.hasNext").value(false))
-                // 첫 번째 북마크 DTO 전체 필드 검증 (최신순이므로 bookmark2)
-                .andExpect(jsonPath("$.data.bookmarks[0].bookmarkId").value(bookmark2.getId()))
-                .andExpect(jsonPath("$.data.bookmarks[0].postId").value(testPost2.getId()))
-                .andExpect(jsonPath("$.data.bookmarks[0].title").value("테스트 게시글 2"))
-                .andExpect(jsonPath("$.data.bookmarks[0].shortSummary").value("게시글 2의 짧은 요약"))
-                .andExpect(jsonPath("$.data.bookmarks[0].url").value("https://test.com/post/2"))
-                .andExpect(jsonPath("$.data.bookmarks[0].companyName").value("테스트회사"))
-                .andExpect(jsonPath("$.data.bookmarks[0].logoUrl").value("https://test.com/logo.png"))
-                .andExpect(jsonPath("$.data.bookmarks[0].publishedAt").exists())
-                .andExpect(jsonPath("$.data.bookmarks[0].thumbnailUrl").value("https://test.com/thumb2.png"))
-                .andExpect(jsonPath("$.data.bookmarks[0].viewCount").value(0))
-                .andExpect(jsonPath("$.data.bookmarks[0].keywords").isArray())
-                .andExpect(jsonPath("$.data.bookmarks[0].isBookmarked").value(true));
-    }
-
-    @Test
-    @DisplayName("북마크 목록 조회 성공 - 커서 기반 페이징")
-    void getBookmarks_Success_WithCursor() throws Exception {
-        // Given - 여러 개의 북마크 생성
-        Bookmark bookmark1 = Bookmark.create(testUser, testPost1, LocalDateTime.now().minusHours(2));
-        Bookmark bookmark2 = Bookmark.create(testUser, testPost2, LocalDateTime.now().minusHours(1));
-        bookmarkRepository.save(bookmark1);
-        bookmarkRepository.save(bookmark2);
-
-        // When & Then - 첫 페이지 조회
-        mockMvc.perform(get("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .param("size", "1"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.isSuccess").value(true))
-                .andExpect(jsonPath("$.data.bookmarks.length()").value(1))
-                .andExpect(jsonPath("$.data.hasNext").value(true))
-                .andExpect(jsonPath("$.data.lastBookmarkId").exists());
-    }
-
     // ===== 읽은 게시글 조회 테스트 =====
 
     @Test
@@ -607,41 +432,6 @@ class ActivityControllerIntegrationTest extends IntegrationTestBase {
     }
 
     // ===== 통합 시나리오 테스트 =====
-
-    @Test
-    @DisplayName("통합 시나리오 - 북마크 추가 후 조회 후 삭제")
-    void integrationScenario_AddGetDeleteBookmark() throws Exception {
-        // 1. 북마크 추가
-        BookmarkRequest addRequest = new BookmarkRequest(testPost1.getId());
-        mockMvc.perform(post("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(addRequest)))
-                .andExpect(status().isCreated());
-
-        // 2. 북마크 목록 조회
-        mockMvc.perform(get("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .param("size", "20"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.data.bookmarks.length()").value(1))
-                .andExpect(jsonPath("$.data.bookmarks[0].postId").value(testPost1.getId()));
-
-        // 3. 북마크 삭제
-        BookmarkRequest deleteRequest = new BookmarkRequest(testPost1.getId());
-        mockMvc.perform(delete("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(deleteRequest)))
-                .andExpect(status().isOk());
-
-        // 4. 삭제 후 목록 조회 - 빈 목록
-        mockMvc.perform(get("/api/v1/activities/bookmarks")
-                        .header("Authorization", "Bearer " + accessToken)
-                        .param("size", "20"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.data.bookmarks").isEmpty());
-    }
 
     @Test
     @DisplayName("통합 시나리오 - 게시글 읽기 후 읽은 목록 조회")
